Add files via upload
In this R Code file, we explore the International Mathematical Olympiad dataset, published in September 2024. The dataset contains results from the International Mathematical Olympiad (IMO) spanning from 1959 to 2024. It includes various aspects of the competition, such as participating countries, the composition of each team, and the timeline of the event. It also captures both individual and team results, including scores for each problem, individual rankings, and medal counts.

Our main objective is to analyze the performance of countries in the IMO over the past 30 years. We then explore how problem difficulty levels (Easy, Medium, Hard) influence the performance of both high-performing and emerging countries. Specifically, we aim to address the following questions:

1. Which countries have performed exceptionally well over the last 30 years, based on country rankings and gold medals?
2. Which countries show strong potential to achieve higher scores and win more medals in the future?
3. How do problem difficulty levels help differentiate the capabilities of the top-performing and high-potential countries?